ref: separate api differences to separate objects
